chore: refactor webhook helpers
