+++ packages/sdk/src/webhooks/index.ts
@@ -1,0 +1,7 @@
+export * from "./client";
+export * from "./types";
+
+/**
+ * @deprecated LinearWebhooks has been renamed: use LinearWebhookClient instead. This alias will be removed in a future version.
+ */
+export { LinearWebhookClient as LinearWebhooks } from "./client";
